refactor: Edit AgreeTermsAndConditions packaging
